@@ -10,7 +10,7 @@ edit_on_github = True
 [options]
 zip_safe = False
 packages = find:
-python_requires = >=3.7
+python_requires = >=3.8
 setup_requires = setuptools_scm
 install_requires =
     numpy
@@ -24,6 +24,7 @@ install_requires =
     pyccl
     sacc
     fgspectra
+    mflike @ git+https://github.com/simonsobs/lat_mflike@master
 
 [options.package_data]
 soliket = *.yaml,*.bibtex,clusters/data/*,clusters/data/selFn_equD56/*,lensing/data/*.txt,lensing/data/*.fits,mflike/*.yaml,tests/*.yaml,data/xcorr_simulated/*.txt,data/CosmoPower/CP_paper/CMB/*.pkl
